+++ app/src/components/administration/Instructions/index.tsx
@@ -19,11 +19,11 @@ import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
 import { RiAddLine, RiPencilLine } from "react-icons/ri";
 import { AiOutlineDownload } from "react-icons/all";
 import { useNavigate } from "react-router-dom";
-import { LoadingSpinner } from "../../components/Loading";
-import { GenericModal } from "../../components/Modal/GenericModal";
-import { api } from "../../lib/axios";
+import { LoadingSpinner } from "../../Loading";
+import { GenericModal } from "../../Modal/GenericModal";
+import { api } from "../../../lib/axios";
 import { toast } from "react-toastify";
-import { Input } from "../../components/admin/Input";
+import { Input } from "../../admin/Input";
 import * as pdfMake from "pdfmake/build/pdfmake";
 import * as pdfFonts from "pdfmake/build/vfs_fonts";
 import { TDocumentDefinitions } from "pdfmake/interfaces";
