+++ app/src/pages/AdminDashboard/instructions.tsx
@@ -2,7 +2,7 @@ import { ChakraProvider, Flex } from "@chakra-ui/react";
 import { Header } from "../../components/admin/Header";
 import { Sidebar } from "../../components/admin/Sidebar";
 import { AdminContainer } from "../AdminDashboard/style";
-import { Instructions } from "../../components/Instructions";
+import { Instructions } from "../../components/administration/Instructions";
 
 export function InstructionsList() {
   return (
